Python/PanOS/requirements.txt
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
future==1.0.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
netmiko==4.5.0
ntc_templates==7.8.0
pan-os-python==1.12.2
pan-python==0.17.0
paramiko==3.5.1
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
setuptools==80.4.0
six==1.17.0
tabulate==0.9.0
textfsm==1.1.3
urllib3==2.4.0
